requirements.txt
aiohttp>=3.8
aiofiles>=23.0
PyYAML>=6.0
tenacity>=8.0
tqdm>=4.60